Add AppDataUpdateServiceMixin and AppDataPreferenceService.
Experiment with a mixin to perform all the updates of instance variables associated with AppData and an extension that will eventually provide the shared_preference interface.
