+++ .idea/modules.xml
@@ -3,6 +3,7 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/kar_kam_1.iml" filepath="$PROJECT_DIR$/kar_kam_1.iml" />
+      <module fileurl="file://$PROJECT_DIR$/android/kar_kam_1_android.iml" filepath="$PROJECT_DIR$/android/kar_kam_1_android.iml" />
     </modules>
   </component>
 </project>
