+++ lib/app_data/app_data_preference_service.dart
@@ -1,0 +1,9 @@
+
+import 'package:kar_kam_1/app_data/app_data.dart';
+import 'package:kar_kam_1/app_data/app_data_update_service_mixin.dart';
+
+class AppDataPreferenceService extends AppData with AppDataUpdateServiceMixin {
+ AppDataPreferenceService() {
+   initialise();
+ }
+}
