+++ lib/app_data/get_it_service.dart
@@ -1,4 +1,4 @@
-
+//  Import flutter packages.
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
