+++ lib/main/kar_kam.dart
@@ -1,0 +1,26 @@
+//  Import flutter packages.
+import 'package:flutter/material.dart';
+
+// Import project-specific files.
+import 'package:kar_kam_1/app_data/app_data.dart';
+import 'package:kar_kam_1/app_data/get_it_service.dart';
+
+class KarKam extends StatelessWidget {
+  const KarKam({super.key});
+
+  // This widget is the root of your application.
+  @override
+  Widget build(BuildContext context) {
+    print('test = ${GetItService.instance<AppData>().test}');
+
+    return MaterialApp(
+      title: 'Kar Kam',
+      theme: ThemeData(
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+        useMaterial3: true,
+      ),
+      home: const Placeholder(),
+    );
+  }
+}
+
